pyproject.toml
[tool.poetry]
name = "contactosapi"
version = "0.1.0"
description = "Backend para la aplicación de Contactos"
authors = ["Roberto <rbarrero@posteo.net>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "*"
djangorestframework = "*"
markdown = "*"
django-extensions = "*"
mysqlclient = "*"
openpyxl = "*"
reportlab = "*"
peewee = "*"
pygments = "*"
django-cors-headers = "*"
PyYAML = "^5.4.1"
uritemplate = "^3.0.1"
iteration-utilities = "^0.11.0"
PyMySQL = "^1.0.2"
django-visago = "^0.2"
django-environ = "^0.4.5"
PyJWT = "^2.1.0"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "*"
autopep8 = "*"
werkzeug = "*"
peewee = "*"
pylint = "*"
pymysql = "*"
black = {version = "^21.6b0", allow-prereleases = true}
coverage = "^5.5"
pylint-django = "^2.4.4"
flake8 = "^3.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "gobex"
url = "https://gitlab.gobex.pri/api/v4/projects/744/packages/pypi/simple"
